Add trained PCAModel for CXR outlier detection.
The model was build with the "CXR Outlier Detector Training" notebook.
The training data was 500 high quality DICOM CXR images from the
TBPortals data. Where HQ was defined as having a modality of CR|XC|RG
and the pixel type was detected as 16-bit integer. 25 components were
selected for the PCA based on comparison of residuals for this
training set, compare to a second set of 500 images from the same HQ
pool the validation set.