Refactor PCAModel with base class: Model
The PCAModel class is a specific instance of a model representation,
and an outlier classifier. The base class: Model, provides an
abstraction for this type of class. A general load_model class method
is provided in Model, which can load any class derived from Model. The
Model class also provides the common functionality for registering an
image to the atlas and transforming images to a reference image
coordinates.