Update to sckit learn 1.3
Migrated the save pickled model by loading, validating and saving with
the newer version of scikit learn. Validation was performed with 400
CXR images from the NIH clinical center dataset. The PCA residuals
where compared to be with in ~0.001 with the old version and 1.3.0 of
scikit learn. The resulting outlier classification matched for all
test cases.